Hot reload split
